+++ .github/workflows/ci.yaml
@@ -42,76 +42,6 @@ jobs:
           token: ${{ secrets.CODESPEED_TOKEN }}
           run: uv run pytest tests/bench_pipeline.py tests/bench_pooled.py --codspeed -n 2 --no-cov -p no:rerunfailures
 
-  memory-tests:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      pull-requests: write
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-        with:
-          # the base commit has to be reachable for the comparison below
-          fetch-depth: 0
-
-      - uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97 # 7.0.0
-        with:
-          python-version: "3.12"
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v7
-        with:
-          enable-cache: true
-          cache-dependency-glob: "uv.lock"
-
-      - name: Install dependencies
-        run: uv sync --group dev
-
-      # peak RSS per pooled config, one spawned process each; reported, never asserted
-      - name: Measure this ref
-        run: MLF_MEM_JSON=mem_pr.json uv run pytest tests/mem_pooled.py --no-cov -s
-
-      # mlforecast is installed editable, so swapping the package source swaps the
-      # library under an unchanged harness -- which is what makes the two runs
-      # comparable, and works even when the base has no tests/mem_pooled.py
-      - name: Measure the base ref
-        if: github.event_name == 'pull_request'
-        run: |
-          base=$(git merge-base ${{ github.event.pull_request.base.sha }} HEAD)
-          echo "comparing against $base"
-          git checkout "$base" -- mlforecast/
-          # a base too old for this harness must not redden the PR
-          MLF_MEM_JSON=mem_base.json uv run pytest tests/mem_pooled.py --no-cov -s || true
-          git checkout HEAD -- mlforecast/
-
-      - name: Compare
-        if: github.event_name == 'pull_request'
-        run: uv run python tests/mem_compare.py mem_base.json mem_pr.json > report.md
-
-      # a fork PR's token is read-only, so this step is allowed to fail; the report
-      # is in the job summary either way
-      - name: Comment on the PR
-        if: github.event_name == 'pull_request'
-        continue-on-error: true
-        uses: actions/github-script@3a2844b7e9c422d3c10d287c895573f7108da1b3 # v9.0.0
-        with:
-          script: |
-            const fs = require('fs');
-            const body = fs.readFileSync('report.md', 'utf8');
-            const marker = '<!-- pooled-memory-report -->';
-            const {owner, repo} = context.repo;
-            const issue_number = context.issue.number;
-            const comments = await github.paginate(
-              github.rest.issues.listComments, {owner, repo, issue_number}
-            );
-            const existing = comments.find(c => c.body.includes(marker));
-            if (existing) {
-              await github.rest.issues.updateComment(
-                {owner, repo, comment_id: existing.id, body}
-              );
-            } else {
-              await github.rest.issues.createComment({owner, repo, issue_number, body});
-            }
-
   performance-tests:
     runs-on: ubuntu-latest
     steps:
